+++ skbio/util/tests/test_misc.py
@@ -10,6 +10,7 @@ import io
 import unittest
 
 import numpy as np
+import numpy.testing as npt
 
 from skbio.util import cardinal_to_ordinal, safe_md5, find_duplicates, get_rng
 from skbio.util._misc import MiniRegistry, chunk_str, resolve_key
@@ -233,46 +234,88 @@ class TestFindDuplicates(unittest.TestCase):
         self.assertEqual(find_duplicates(gen()), set(['a', 2]))
 
 
+
 class TestGetRng(unittest.TestCase):
 
     def test_get_rng(self):
 
-        # no seed
-        obs0 = get_rng()
-        self.assertTrue(isinstance(obs0, np.random.Generator))
+        # returns random generator
+        res = get_rng()
+        self.assertIsInstance(res, np.random.RandomState)
 
-        # integer seed
-        obs1 = get_rng(42)
-        self.assertTrue(isinstance(obs1, np.random.Generator))
+        # seed is Python integer
+        res = get_rng(42)
+        self.assertIsInstance(res, np.random.Generator)
+        obs = np.array([res.integers(100) for _ in range(5)])
+        exp = np.array([8, 77, 65, 43, 43])
+        npt.assert_array_equal(obs, exp)
 
-        # generator instance
-        obs2 = get_rng(obs1)
-        self.assertTrue(isinstance(obs2, np.random.Generator))
+        # seed is NumPy integer
+        res = get_rng(np.uint8(42))
+        self.assertIsInstance(res, np.random.Generator)
+        obs = np.array([res.integers(100) for _ in range(5)])
+        npt.assert_array_equal(obs, exp)
 
-        # invalide seed
-        msg = ('Invalid seed. It must be an integer or an instance of '
-               'np.random.Generator.')
-        with self.assertRaises(ValueError) as cm:
-            get_rng('hello')
-        self.assertEqual(str(cm.exception), msg)
+        # seed is new Generator
+        res = get_rng(res)
+        self.assertIsInstance(res, np.random.Generator)
+        obs = np.array([res.integers(100) for _ in range(5)])
+        exp = np.array([85, 8, 69, 20, 9])
+        npt.assert_array_equal(obs, exp)
 
-        # test if seeds are disjoint and results are reproducible
+        # test if integer seeds are disjoint
         obs = [get_rng(i).integers(1e6) for i in range(10)]
         exp = [850624, 473188, 837575, 811504, 726442,
                670790, 445045, 944904, 719549, 421547]
         self.assertListEqual(obs, exp)
 
-        # mimic legacy numpy
-        delattr(np.random, 'default_rng')
-        delattr(np.random, 'Generator')
-        msg = ('The installed NumPy version does not support '
-               'random.Generator. Please use NumPy >= 1.17.')
+        # no seed: use current random state
+        np.random.seed(42)
+        res = get_rng()
+        self.assertIsInstance(res, np.random.RandomState)
+        obs = np.array([res.randint(100) for _ in range(5)])
+        exp = np.array([51, 92, 14, 71, 60])
+        npt.assert_array_equal(obs, exp)
+
+        # reset random state to reproduce output
+        np.random.seed(42)
+        res = get_rng()
+        obs = np.array([res.randint(100) for _ in range(5)])
+        npt.assert_array_equal(obs, exp)
+
+        # seed is legacy RandomState
+        res = get_rng(np.random.RandomState(42))
+        obs = np.array([res.randint(100) for _ in range(5)])
+        npt.assert_array_equal(obs, exp)
+
+        # test if legacy random states are disjoint
+        obs = [get_rng(np.random.RandomState(i)).randint(1e6) for i in range(5)]
+        exp = [985772, 128037, 875688, 71530, 991406]
+        self.assertListEqual(obs, exp)
+
+        # invalid seed
+        msg = "Invalid seed. It must be an integer or a random generator instance."
         with self.assertRaises(ValueError) as cm:
-            get_rng()
+            get_rng("hello")
         self.assertEqual(str(cm.exception), msg)
+
+        # mimic legacy NumPy (no Generator)
+        delattr(np.random, "Generator")
+        res = get_rng(42)
+        self.assertIsInstance(res, np.random.RandomState)
+
+        # mimic unknown NumPy (neither Generator nor RandomState)
+        delattr(np.random, "RandomState")
+        msg = "Supported random generators are not available."
         with self.assertRaises(ValueError) as cm:
-            get_rng('hello')
+            get_rng(42)
         self.assertEqual(str(cm.exception), msg)
+
+        # mimic future NumPy (no RandomState)
+        setattr(np.random, "Generator", int)
+        setattr(np.random, "default_rng", lambda: 1)
+        res = get_rng()
+        self.assertEqual(res, 1)
 
 
 if __name__ == '__main__':